+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Release Notes
 =============
 
 - :release:`unreleased <????-??-??>`
+- :feature:`1336` In the CfP settings, a warning will be shown next to the deadline setting if some session types have different deadlines.
 - :feature:`1324` Session and speaker images will now be shown in an overlay on click, instead of opening a new page.
 - :feature:`1350` pretalx will now warn you if you have empty URLs in email templates (which can easily happen with Markdown by accident).
 - :feature:`1244` Emails can now be sent both per-speaker and per-proposal, though some placeholders are not available in per-speaker emails (such as the proposal title).